Update script.js for YouTube Emotes
Updating this part allows YouTube emotes to be repeated and not just return text. It also gets rid of the extra colons that appear since the emotes have them in before and after the emote ID

I am not as knowledgeable with coding so please make sure I didn't miss something, but it works well for me.